requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
psycopg2
email-validator
python-dotenv
faker
locust